+++ currency/def.go
@@ -125,6 +125,18 @@ func (d *Def) Zero() num.Amount {
 	return num.MakeAmount(0, d.Subunits)
 }
 
+// Rescale takes the provided amount and ensures its scale matches
+// that of the currency.
+func (d *Def) Rescale(a num.Amount) num.Amount {
+	return a.Rescale(d.Subunits)
+}
+
+// RescaleUp ensures tha the amount has *at least* the same
+// precision as the currency.
+func (d *Def) RescaleUp(a num.Amount) num.Amount {
+	return a.RescaleUp(d.Subunits)
+}
+
 // Definitions provides an array of all currency definitions
 // ordered by priority.
 func Definitions() []*Def {